Plugin engine using Rhai scripting language
This is a proof-of-concept integration of the "Rhai" language as a
plugin / scripting language.

Any script ending with ".rhai" located in the directory
"~/.config/helix/plugins/" is automatically loaded when helix is
started.

If the script defines an optional "on_start" function it will be called
on startup. Likewise, optional "on_resize", "on_key_press" and
"on_mouse_event" functions are called on those respective events.